move Triple struct para models
