+++ api/Cargo.toml
@@ -2,6 +2,7 @@
 name = "api"
 version = "0.1.0"
 edition = "2024"
+default-run = "grpg"
 
 [dependencies]
 actix-cors = "0.7.1"
@@ -19,3 +20,7 @@ utoipa-scalar = { version = "0.3.0", features = ["actix-web"] }
 sea-orm = { version = "^0.12.0", features = [ "sqlx-sqlite", "runtime-async-std-native-tls", "macros" ] }
 futures = "0.3.28"
 serde_json = "1.0.141"
+
+[[bin]]
+name = "grpg"
+path = "src/main.rs"
